backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0
python-dotenv>=1.0.0
supabase>=2.3.0
tensorflow>=2.15.0
pillow>=10.0.0
numpy>=1.24.0
